Select size: CreateQuiz reorganization
